+++ packages/adapter-netlify/src/edge.js
@@ -3,18 +3,37 @@ import { manifest } from 'MANIFEST';
 
 const server = new Server(manifest);
 
+/**
+ * We don't know the origin until we receive a request, but
+ * that's guaranteed to happen before we call `read`
+ * @type {string}
+ */
+let origin;
+
 const initialized = server.init({
 	// @ts-ignore
-	env: Deno.env.toObject()
+	env: Deno.env.toObject(),
+	read: async (file) => {
+		const url = `${origin}/${file}`;
+		const response = await fetch(url);
+
+		if (!response.ok) {
+			throw new Error(
+				`read(...) failed: could not fetch ${url} (${response.status} ${response.statusText})`
+			);
+		}
+
+		return response.body;
+	}
 });
 
-/**
- * @param { Request } request
- * @param { any } context
- * @returns { Promise<Response> }
- */
+/** @type {import('@netlify/edge-functions').EdgeFunction} */
 export default async function handler(request, context) {
-	await initialized;
+	if (!origin) {
+		origin = new URL(request.url).origin;
+		await initialized;
+	}
+
 	return server.respond(request, {
 		platform: { context },
 		getClientAddress() {