+++ packages/kit/kit.vitest.config.js
@@ -1,20 +1,22 @@
+import { fileURLToPath } from 'node:url';
 import { defineConfig } from 'vitest/config';
 
 // this file needs a custom name so that the numerous test subprojects don't all pick it up
 export default defineConfig({
+	define: {
+		__SVELTEKIT_SERVER_TRACING_ENABLED__: false
+	},
 	server: {
 		watch: {
 			ignored: ['**/node_modules/**', '**/.svelte-kit/**']
 		}
 	},
 	test: {
-		// shave a couple seconds off the tests
-		isolate: false,
-		poolOptions: {
-			threads: {
-				singleThread: true
-			}
+		alias: {
+			'__sveltekit/paths': fileURLToPath(new URL('./test/mocks/path.js', import.meta.url))
 		},
+		pool: 'threads',
+		maxWorkers: 1,
 		include: ['src/**/*.spec.js'],
 		exclude: [
 			'**/node_modules/**',