+++ packages/kit/src/core/generate_manifest/find_server_assets.js
@@ -13,7 +13,6 @@ export function find_server_assets(build_data, routes) {
 	 */
 	const used_nodes = new Set([0, 1]);
 
-	// TODO add hooks.server.js asset imports
 	/** @type {Set<string>} */
 	const server_assets = new Set();
 
@@ -41,11 +40,16 @@ export function find_server_assets(build_data, routes) {
 
 	for (const n of used_nodes) {
 		const node = build_data.manifest_data.nodes[n];
+		if (node?.universal) add_assets(node.universal);
 		if (node?.server) add_assets(node.server);
 	}
 
 	if (build_data.manifest_data.hooks.server) {
 		add_assets(build_data.manifest_data.hooks.server);
+	}
+
+	if (build_data.manifest_data.hooks.universal) {
+		add_assets(build_data.manifest_data.hooks.universal);
 	}
 
 	return Array.from(server_assets);