+++ packages/kit/src/utils/env.js
@@ -1,33 +1,13 @@
 /**
  * @param {Record<string, string>} env
- * @param {{
- * 		public_prefix: string;
- * 		private_prefix: string;
- * }} prefixes
+ * @param {string} allowed
+ * @param {string} disallowed
  * @returns {Record<string, string>}
  */
-export function filter_private_env(env, { public_prefix, private_prefix }) {
+export function filter_env(env, allowed, disallowed) {
 	return Object.fromEntries(
 		Object.entries(env).filter(
-			([k]) =>
-				k.startsWith(private_prefix) && (public_prefix === '' || !k.startsWith(public_prefix))
-		)
-	);
-}
-
-/**
- * @param {Record<string, string>} env
- * @param {{
- * 		public_prefix: string;
- *    private_prefix: string;
- * }} prefixes
- * @returns {Record<string, string>}
- */
-export function filter_public_env(env, { public_prefix, private_prefix }) {
-	return Object.fromEntries(
-		Object.entries(env).filter(
-			([k]) =>
-				k.startsWith(public_prefix) && (private_prefix === '' || !k.startsWith(private_prefix))
+			([k]) => k.startsWith(allowed) && (disallowed === '' || !k.startsWith(disallowed))
 		)
 	);
 }