+++ packages/kit/test/apps/async/playwright.config.js
@@ -1,0 +1,11 @@
+import process from 'node:process';
+import { config } from '../../utils.js';
+import { defineConfig } from '@playwright/test';
+
+export default defineConfig({
+	...config,
+	webServer: {
+		command: process.env.DEV ? `pnpm dev` : `pnpm build && pnpm preview`,
+		port: process.env.DEV ? 5173 : 4173
+	}
+});