+++ packages/kit/test/apps/async/src/routes/remote/+page.js
@@ -1,0 +1,7 @@
+import { echo } from './query-command.remote';
+
+export async function load() {
+	return {
+		echo_result: await echo('Hello world')
+	};
+}