+++ packages/kit/test/apps/async/src/routes/remote/query-redirect/from-common-layout/+layout.svelte
@@ -1,0 +1,12 @@
+<script>
+	import { page } from '$app/state';
+	import { layoutRedirect } from '../redirect.remote';
+
+	let { children } = $props();
+</script>
+
+<p id="layout-query">
+	on page {await layoutRedirect(page.url.pathname)} (== {page.url.pathname})
+</p>
+
+{@render children()}