+++ packages/kit/test/apps/basics/src/routes/actions/enhance/+page.svelte
@@ -37,7 +37,9 @@
 	use:enhance={({ controller }) => {
 		previous?.abort();
 		previous = controller;
-		return () => count++;
+		return () => {
+			count++;
+		};
 	}}
 >
 	<button class="form2">Submit</button>