+++ packages/kit/test/apps/basics/src/routes/prerendering/prerendered-endpoint/proxy/+server.js
@@ -1,3 +1,7 @@
-export async function GET({ fetch }) {
+export async function GET({ fetch, url }) {
+	if (url.searchParams.get('api-with-param-option') === 'prerendered') {
+		return await fetch('/prerendering/prerendered-endpoint/api-with-param/prerendered');
+	}
+
 	return await fetch('/prerendering/prerendered-endpoint/api');
 }