+++ packages/kit/test/apps/dev-only/src/routes/headers/invalid/+server.js
@@ -1,0 +1,9 @@
+/** @type {import("@sveltejs/kit").RequestHandler} */
+export function GET({ setHeaders }) {
+	setHeaders({
+		'cache-control': 'totally-invalid',
+		'content-type': 'not-a-real-type'
+	});
+
+	return new Response('Testing invalid headers');
+}