+++ packages/kit/test/apps/options/svelte.config.js
@@ -1,3 +1,5 @@
+import process from 'node:process';
+
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	extensions: ['.jesuslivesineveryone', '.whokilledthemuffinman', '.svelte.md', '.svelte'],
@@ -21,7 +23,7 @@ const config = {
 			// while we specify a path for the service worker, we expect it to not exist in the test
 			serviceWorker: 'source/service-worker'
 		},
-		appDir: '_wheee',
+		appDir: '_wheee/nested',
 		inlineStyleThreshold: 1024,
 		outDir: '.custom-out-dir',
 		output: {
@@ -29,12 +31,21 @@ const config = {
 		},
 		paths: {
 			base: '/path-base',
-			assets: 'https://cdn.example.com/stuff'
+			// @ts-expect-error our env var string can't match the https template literal
+			assets: process.env.PATHS_ASSETS || ''
 		},
 		env: {
 			dir: './env-dir',
 			publicPrefix: 'GO_AWAY_',
 			privatePrefix: 'TOP_SECRET_SHH'
+		},
+		router: {
+			resolution: /** @type {'client' | 'server'} */ (process.env.ROUTER_RESOLUTION) || 'client'
+		}
+	},
+	compilerOptions: {
+		experimental: {
+			async: true
 		}
 	}
 };