+++ packages/kit/test/apps/writes/test/test.js
@@ -53,4 +53,63 @@ test.describe('Filesystem updates', () => {
 			fs.writeFileSync(file, contents.replace(/PLACEHOLDER:\d+/, 'PLACEHOLDER:0'));
 		}
 	});
+
+	test('Universal node is updated when page options change', async ({
+		page,
+		javaScriptEnabled,
+		get_computed_style
+	}) => {
+		test.skip(!process.env.DEV || !javaScriptEnabled);
+
+		const file = fileURLToPath(new URL('../src/routes/universal/+page.js', import.meta.url));
+		const contents = fs.readFileSync(file, 'utf-8');
+
+		try {
+			fs.writeFileSync(file, contents.replace(/export const ssr = false;\n/, ''));
+			await page.goto('/universal', { wait_for_started: false });
+			expect(await get_computed_style('body', 'background-color')).not.toBe('rgb(255, 0, 0)');
+			await expect(page.locator('h1')).toHaveText('Internal Error');
+		} finally {
+			fs.writeFileSync(file, contents);
+		}
+
+		await page.waitForTimeout(500); // this is the rare time we actually need waitForTimeout; we have no visibility into whether the module graph has been invalidated
+		// a reload is required because Vite HMR doesn't trigger if the page has never loaded successfully
+		await page.reload();
+		expect(await get_computed_style('body', 'background-color')).toBe('rgb(255, 0, 0)');
+
+		try {
+			fs.writeFileSync(file, contents.replace(/export const ssr = .*;/, 'export const ssr = !1;'));
+			await page.waitForTimeout(500); // this is the rare time we actually need waitForTimeout; we have no visibility into whether the module graph has been invalidated
+			expect(await get_computed_style('body', 'background-color')).not.toBe('rgb(255, 0, 0)');
+			await expect(page.locator('h1')).toHaveText('Internal Error');
+		} finally {
+			fs.writeFileSync(file, contents.replace(/\\nexport const ssr = false;\\n/, ''));
+		}
+	});
+
+	test('Universal node is updated when parent page options change', async ({
+		page,
+		javaScriptEnabled,
+		get_computed_style
+	}) => {
+		test.skip(!process.env.DEV || !javaScriptEnabled);
+
+		const file = fileURLToPath(
+			new URL('../src/routes/universal/parent-changed/+layout.js', import.meta.url)
+		);
+		const contents = fs.readFileSync(file, 'utf-8');
+
+		try {
+			await page.goto('/universal/parent-changed');
+			expect(await get_computed_style('body', 'background-color')).toBe('rgb(255, 0, 0)');
+
+			fs.writeFileSync(file, contents.replace(/export const ssr = false;/, ''));
+			await page.waitForTimeout(500); // this is the rare time we actually need waitForTimeout; we have no visibility into whether the module graph has been invalidated
+			expect(await get_computed_style('body', 'background-color')).not.toBe('rgb(255, 0, 0)');
+			await expect(page.locator('h1')).toHaveText('Internal Error');
+		} finally {
+			fs.writeFileSync(file, contents);
+		}
+	});
 });